refactor(internal): move HTTP/2 and unified adapter headers to internal
Move HTTP/2 protocol implementation headers and unified adapter headers
from public include directory to internal location. This is part of
the facade pattern migration (Epic #577) to reduce public API surface.

Changes:
- Move 6 HTTP/2 headers to src/internal/protocols/http2/
- Move 6 unified adapter headers to src/internal/unified/adapters/
- Create deprecated redirect headers in original locations
- Update include paths in all source files

Part of #577
Closes #634
