+++ src/protocol/tcp.cpp
@@ -31,8 +31,8 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *****************************************************************************/
 
 #include "kcenon/network/protocol/tcp.h"
-#include "kcenon/network/unified/adapters/tcp_connection_adapter.h"
-#include "kcenon/network/unified/adapters/tcp_listener_adapter.h"
+#include "internal/unified/adapters/tcp_connection_adapter.h"
+#include "internal/unified/adapters/tcp_listener_adapter.h"
 
 #include <chrono>
 #include <random>
