+++ src/protocols/http2/hpack.cpp
@@ -30,7 +30,7 @@ OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *****************************************************************************/
 
-#include "kcenon/network/protocols/http2/hpack.h"
+#include "internal/protocols/http2/hpack.h"
 #include <algorithm>
 #include <cstring>
