+++ src/protocols/http2/http2_client.cpp
@@ -30,7 +30,7 @@ OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *****************************************************************************/
 
-#include "kcenon/network/protocols/http2/http2_client.h"
+#include "internal/protocols/http2/http2_client.h"
 
 #include "kcenon/network/tracing/span.h"
 #include "kcenon/network/tracing/trace_context.h"
