+++ src/protocols/http2/http2_server_stream.cpp
@@ -30,7 +30,7 @@ OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *****************************************************************************/
 
-#include "kcenon/network/protocols/http2/http2_server_stream.h"
+#include "internal/protocols/http2/http2_server_stream.h"
 
 namespace kcenon::network::protocols::http2
 {
